minipc-client: add a flag MINIPC_FLAG_MSG_NOSIGNAL
The problem that I hit is when the ipc server goes down, there is no way
to check (using minipc) whether server is down. When minipc_call is issued
on a broken connection the client is getting SIGPIPE signal. It is more
convenient to check the return value of minipc_call, instead of handling
the signal.
This commit introcudes such choice.

Signed-off-by: Adam Wujek <adam.wujek@cern.ch>