all: relicensing to LGPLv2.1 to fix compatibility with ppsi
After a discussion with CERN, it has been agreed to allow a relicensing
of minipc source code to LGPL v2.1 in order to allow ppsi to be LGPL
v2.1 wich it is how it was designed.
The source code of minipc is fully under CERN copyright with some fixes
from Seven Solutions S.L. and all parties agree for this relicensing.